Refactored config variables
