+++ tsdat/config/variables/ureg.py
@@ -1,0 +1,4 @@
+from pint import UnitRegistry
+
+ureg = UnitRegistry()
+ureg.define("unitless = count = 1")  # type: ignore
